Module serenity_voice_model::payload

Message bodies used in gateway event-handling.

Structs

ClientConnect

Message indicating that another user has connected to the voice channel.

ClientDisconnect

Message indicating that another user has disconnected from the voice channel.

Heartbeat

Used to keep the websocket connection alive.

HeartbeatAck

Heartbeat ACK, received by the client to show the server’s receipt of a heartbeat.

Hello

Used to determine how often the client must send a heartbeat.

Identify

Used to begin a voice websocket connection.

Ready

RTP server’s connection offer and supported encryption modes.

Resume

Sent by the client after a disconnect to attempt to resume a session.

SelectProtocol

Used to select the voice protocol and encryption mechanism.

SessionDescription

Server’s confirmation of a negotiated encryption scheme.

Speaking

Used to indicate which users are speaking, or to inform Discord that the client is now speaking.